+++ test/lean_test.dart
@@ -20,7 +20,7 @@ void main() {
     group('connect', () {
       test('partial params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=connect&customer_id=dda80d32-4062-404c-abe7-ba9b9290c873&permissions=identity&permissions=accounts&permissions=balance&permissions=transactions&permissions=payments';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=connect&customer_id=dda80d32-4062-404c-abe7-ba9b9290c873&permissions=identity&permissions=accounts&permissions=balance&permissions=transactions&permissions=payments';
 
         final initializationURL = leanSdk.connect(
           customerId: customerId,
@@ -44,7 +44,7 @@ void main() {
 
       test('all params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=connect&customer_id=dda80d32-4062-404c-abe7-ba9b9290c873&permissions=identity&permissions=accounts&permissions=balance&permissions=transactions&permissions=payments&bank_identifier=LEANMB1_SAU&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&access_to=10-10-2023&access_from=10-05-2023&access_token=test&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=connect&customer_id=dda80d32-4062-404c-abe7-ba9b9290c873&permissions=identity&permissions=accounts&permissions=balance&permissions=transactions&permissions=payments&bank_identifier=LEANMB1_SAU&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&access_to=10-10-2023&access_from=10-05-2023&access_token=test&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success&destination_alias=TestingCo&destination_avatar=https://dev.leantech.me/success.png';
 
         final initializationURL = leanSdk.connect(
             customerId: customerId,
@@ -61,7 +61,9 @@ void main() {
             bankIdentifier: 'LEANMB1_SAU',
             failRedirectUrl: 'https://dev.leantech.me/fail',
             successRedirectUrl: 'https://dev.leantech.me/success',
-            paymentDestinationId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0');
+            paymentDestinationId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            destinationAlias: 'TestingCo',
+            destinationAvatar: 'https://dev.leantech.me/success.png');
 
         expect(initializationURL, equals(expectedUrl));
       });
@@ -70,7 +72,7 @@ void main() {
     group('reconnect', () {
       test('partial params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=reconnect&reconnect_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=reconnect&reconnect_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
 
         final initializationURL = leanSdk.reconnect(
           reconnectId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
@@ -81,12 +83,13 @@ void main() {
 
       test('all params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=reconnect&reconnect_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&access_token=test';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=reconnect&reconnect_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&access_token=test&destination_alias=TestingCo&destination_avatar=https://dev.leantech.me/success.png';
 
         final initializationURL = leanSdk.reconnect(
-          reconnectId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
-          accessToken: 'test',
-        );
+            reconnectId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            accessToken: 'test',
+            destinationAlias: 'TestingCo',
+            destinationAvatar: 'https://dev.leantech.me/success.png');
 
         expect(initializationURL, equals(expectedUrl));
       });
@@ -95,7 +98,7 @@ void main() {
     group('createBeneficiary', () {
       test('partial params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=createBeneficiary&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=createBeneficiary&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
 
         final initializationURL = leanSdk.createBeneficiary(
           customerId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
@@ -106,17 +109,18 @@ void main() {
 
       test('all params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=createBeneficiary&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&payment_source_id=8b3b7960-c4a1-41da-8ad0-5df36cf67540&access_token=test&entity_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=createBeneficiary&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&payment_source_id=8b3b7960-c4a1-41da-8ad0-5df36cf67540&access_token=test&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&entity_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&destination_alias=TestingCo&destination_avatar=https://dev.leantech.me/success.png';
 
         final initializationURL = leanSdk.createBeneficiary(
-          accessToken: 'test',
-          entityId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
-          customerId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
-          paymentSourceId: '8b3b7960-c4a1-41da-8ad0-5df36cf67540',
-          failRedirectUrl: 'https://dev.leantech.me/fail',
-          successRedirectUrl: 'https://dev.leantech.me/success',
-          paymentDestinationId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
-        );
+            accessToken: 'test',
+            entityId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            customerId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            paymentSourceId: '8b3b7960-c4a1-41da-8ad0-5df36cf67540',
+            failRedirectUrl: 'https://dev.leantech.me/fail',
+            successRedirectUrl: 'https://dev.leantech.me/success',
+            paymentDestinationId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            destinationAlias: 'TestingCo',
+            destinationAvatar: 'https://dev.leantech.me/success.png');
 
         expect(initializationURL, equals(expectedUrl));
       });
@@ -125,7 +129,7 @@ void main() {
     group('createPaymentSource', () {
       test('partial params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=createPaymentSource&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=createPaymentSource&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
 
         final initializationURL = leanSdk.createPaymentSource(
           customerId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
@@ -136,16 +140,17 @@ void main() {
 
       test('all params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=createPaymentSource&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&bank_identifier=LEANMB1_SAU&access_token=test&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=createPaymentSource&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&bank_identifier=LEANMB1_SAU&access_token=test&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success&destination_alias=TestingCo&destination_avatar=https://dev.leantech.me/success.png';
 
         final initializationURL = leanSdk.createPaymentSource(
-          accessToken: 'test',
-          bankIdentifier: 'LEANMB1_SAU',
-          customerId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
-          failRedirectUrl: 'https://dev.leantech.me/fail',
-          successRedirectUrl: 'https://dev.leantech.me/success',
-          paymentDestinationId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
-        );
+            accessToken: 'test',
+            bankIdentifier: 'LEANMB1_SAU',
+            customerId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            failRedirectUrl: 'https://dev.leantech.me/fail',
+            successRedirectUrl: 'https://dev.leantech.me/success',
+            paymentDestinationId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            destinationAlias: 'TestingCo',
+            destinationAvatar: 'https://dev.leantech.me/success.png');
 
         expect(initializationURL, equals(expectedUrl));
       });
@@ -154,7 +159,7 @@ void main() {
     group('updatePaymentSource', () {
       test('partial params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=updatePaymentSource&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&payment_source_id=8b3b7960-c4a1-41da-8ad0-5df36cf67540&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=updatePaymentSource&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&payment_source_id=8b3b7960-c4a1-41da-8ad0-5df36cf67540&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
 
         final initializationURL = leanSdk.updatePaymentSource(
           customerId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
@@ -167,17 +172,18 @@ void main() {
 
       test('all params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=updatePaymentSource&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&payment_source_id=8b3b7960-c4a1-41da-8ad0-5df36cf67540&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&access_token=test&entity_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=updatePaymentSource&customer_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&payment_source_id=8b3b7960-c4a1-41da-8ad0-5df36cf67540&payment_destination_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&access_token=test&entity_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success&destination_alias=TestingCo&destination_avatar=https://dev.leantech.me/success.png';
 
         final initializationURL = leanSdk.updatePaymentSource(
-          accessToken: 'test',
-          entityId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
-          customerId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
-          paymentSourceId: '8b3b7960-c4a1-41da-8ad0-5df36cf67540',
-          paymentDestinationId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
-          failRedirectUrl: 'https://dev.leantech.me/fail',
-          successRedirectUrl: 'https://dev.leantech.me/success',
-        );
+            accessToken: 'test',
+            entityId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            customerId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            paymentSourceId: '8b3b7960-c4a1-41da-8ad0-5df36cf67540',
+            paymentDestinationId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
+            failRedirectUrl: 'https://dev.leantech.me/fail',
+            successRedirectUrl: 'https://dev.leantech.me/success',
+            destinationAlias: 'TestingCo',
+            destinationAvatar: 'https://dev.leantech.me/success.png');
 
         expect(initializationURL, equals(expectedUrl));
       });
@@ -186,7 +192,7 @@ void main() {
     group('pay', () {
       test('partial params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=pay&payment_intent_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=pay&payment_intent_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0';
 
         final initializationURL = leanSdk.pay(
           paymentIntentId: '617207b3-a4d4-4413-ba1b-b8d32efd58a0',
@@ -197,16 +203,17 @@ void main() {
 
       test('all params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=pay&payment_intent_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&account_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&show_balances=true&access_token=test&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=pay&payment_intent_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&account_id=617207b3-a4d4-4413-ba1b-b8d32efd58a0&show_balances=true&access_token=test&fail_redirect_url=https://dev.leantech.me/fail&success_redirect_url=https://dev.leantech.me/success&destination_alias=TestingCo&destination_avatar=https://dev.leantech.me/success.png';
 
         final initializationURL = leanSdk.pay(
-          accessToken: 'test',
-          paymentIntentId: "617207b3-a4d4-4413-ba1b-b8d32efd58a0",
-          accountId: "617207b3-a4d4-4413-ba1b-b8d32efd58a0",
-          showBalances: true,
-          failRedirectUrl: 'https://dev.leantech.me/fail',
-          successRedirectUrl: 'https://dev.leantech.me/success',
-        );
+            accessToken: 'test',
+            paymentIntentId: "617207b3-a4d4-4413-ba1b-b8d32efd58a0",
+            accountId: "617207b3-a4d4-4413-ba1b-b8d32efd58a0",
+            showBalances: true,
+            failRedirectUrl: 'https://dev.leantech.me/fail',
+            successRedirectUrl: 'https://dev.leantech.me/success',
+            destinationAlias: 'TestingCo',
+            destinationAvatar: 'https://dev.leantech.me/success.png');
 
         expect(initializationURL, equals(expectedUrl));
       });
@@ -215,7 +222,7 @@ void main() {
     group('verifyAddress', () {
       test('partial params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=verifyAddress&customer_id=test-customer-id&customer_name=test-customer-name&permissions=identity';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=verifyAddress&customer_id=test-customer-id&customer_name=test-customer-name&permissions=identity';
 
         final initializationURL = leanSdk.verifyAddress(
             customerId: 'test-customer-id',
@@ -227,14 +234,46 @@ void main() {
 
       test('all params: returns the correct URL', () {
         const expectedUrl =
-            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.8&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=verifyAddress&customer_id=test-customer-id&customer_name=test-customer-name&permissions=identity&access_token=test-access-token';
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=verifyAddress&customer_id=test-customer-id&customer_name=test-customer-name&permissions=identity&access_token=test-access-token&destination_alias=TestingCo&destination_avatar=https://dev.leantech.me/success.png';
 
         final initializationURL = leanSdk.verifyAddress(
-          customerId: 'test-customer-id',
-          customerName: 'test-customer-name',
-          permissions: [LeanPermissions.identity],
-          accessToken: 'test-access-token',
-        );
+            customerId: 'test-customer-id',
+            customerName: 'test-customer-name',
+            permissions: [LeanPermissions.identity],
+            accessToken: 'test-access-token',
+            destinationAlias: 'TestingCo',
+            destinationAvatar: 'https://dev.leantech.me/success.png');
+
+        expect(initializationURL, equals(expectedUrl));
+      });
+    });
+
+    group('authorizeConsent', () {
+      test('partial params: returns the correct URL', () {
+        const expectedUrl =
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=authorizeConsent&customer_id=b2f3537b-2e38-4067-bf61-7398cf7d4934&consent_id=7ebe7449-fd93-4657-be82-fcc3697262c4&fail_redirect_url=https://www.leantech.me/failure&success_redirect_url=https://www.leantech.me/success';
+
+        final initializationURL = leanSdk.authorizeConsent(
+            customerId: 'b2f3537b-2e38-4067-bf61-7398cf7d4934',
+            consentId: '7ebe7449-fd93-4657-be82-fcc3697262c4',
+            failRedirectUrl: 'https://www.leantech.me/failure',
+            successRedirectUrl: 'https://www.leantech.me/success');
+
+        expect(initializationURL, equals(expectedUrl));
+      });
+
+      test('all params: returns the correct URL', () {
+        const expectedUrl =
+            'https://cdn.leantech.me/link/loader/prod/ae/latest/lean-sdk.html?implementation=webview-hosted-html&implementation_config=platform+mobile&implementation_config=sdk+flutter&implementation_config=os+macos&implementation_config=sdk_version+3.0.9&implementation_config=is_version_pinned+false&app_token=9fb9e934-9efb-4e7e-a508-de67c0839be0&sandbox=false&language=en&version=latest&country=ae&env=production&method=authorizeConsent&customer_id=b2f3537b-2e38-4067-bf61-7398cf7d4934&consent_id=7ebe7449-fd93-4657-be82-fcc3697262c4&fail_redirect_url=https://www.leantech.me/failure&success_redirect_url=https://www.leantech.me/success&access_token=test-access-token&destination_alias=TestingCo&destination_avatar=https://dev.leantech.me/success.png';
+
+        final initializationURL = leanSdk.authorizeConsent(
+            customerId: 'b2f3537b-2e38-4067-bf61-7398cf7d4934',
+            consentId: '7ebe7449-fd93-4657-be82-fcc3697262c4',
+            failRedirectUrl: 'https://www.leantech.me/failure',
+            successRedirectUrl: 'https://www.leantech.me/success',
+            accessToken: 'test-access-token',
+            destinationAlias: 'TestingCo',
+            destinationAvatar: 'https://dev.leantech.me/success.png');
 
         expect(initializationURL, equals(expectedUrl));
       });